requirements.txt
# Django核心框架
django==5.2.8

# 后端API（如果需要）
djangorestframework==3.15.2

# 数据库驱动（MySQL）
mysqlclient==2.2.4

# 数据可视化（对应你之前的Echarts项目）
pyecharts==2.0.4

# 工具类库
python-dotenv==1.0.0  # 环境变量管理
pillow==10.3.0        # 图片处理（如果有上传功能）